fix: modify existing code for modularity
