--- lex_main/src/components/Dashboard.jsx
@@ -1,47 +1,0 @@
-/**
- * Dashboard.jsx
- * 
- * The user dashboard
- */
-import React, { useState } from "react";
-import { Link, useHistory } from "react-router-dom";
-import { Card, Button, Alert } from "react-bootstrap";
-import { useAuth } from "../contexts/AuthContext.js"
-
-export default function Dashboard() {
-
-    const { currentUser, logout } = useAuth();                    // Authentication Context
-    const [error, setError] = useState('');         // Error State
-    const history = useHistory();
-    
-
-    async function handleLogout() {
-        setError('');
-        
-        try {
-            await logout();
-            history.push('/')
-        } catch {
-            setError('Failed to log out')
-        }
-    }
-
-    return (
-        <div>
-            <Card>
-                <Card.Body>
-                    <h1 className="logo text-center">LEX</h1>
-                    <h1 className="text-center mb-3 fw-normal h3">Profile</h1>
-                    <strong>Email:</strong> {currentUser.email}
-                    <Link to="/update-profile" className="btn btn-primary w-100 mt-3">Update Profile</Link>
-
-                    {/* <!-- If there is an error, it will render an error message --> */}
-                    {error && <Alert variant="danger">{error}</Alert>}
-                </Card.Body>
-            </Card>
-            <div className="text-center">
-                <Button variant="link" onClick={handleLogout}>Log Out</Button>
-            </div>
-        </div>
-    )
-}
