+++ lex_main/src/components/authentication/Login.jsx
@@ -6,11 +6,11 @@
 
 import React from "react";
 import { Link } from "react-router-dom";
-import { Container } from "react-bootstrap";
 
 import "../../index.css";
 import Footer from "../Footer";
 import LoginForm from "./LoginForm";
+import CenteredContainer from "./CenteredContainer";
 
 function Login() {
   /**
@@ -19,18 +19,13 @@ function Login() {
 
   return (
     <div className="gradient-background">
-      <Container
-        className="d-flex align-items-center justify-content-center"
-        style={{ minHeight: "100vh" }}
-      >
-        <div className="w-100" style={{ maxWidth: "400px" }}>
-          <Link to="/landing" style={{ textDecoration: "none" }}>
-            <h1 className="logo logo-big logo-white text-center">LEX</h1>
-          </Link>
-          <LoginForm />
-          <Footer theme="light" />
-        </div>
-      </Container>
+      <CenteredContainer>
+        <Link to="/landing" style={{ textDecoration: "none" }}>
+          <h1 className="logo logo-big logo-white text-center">LEX</h1>
+        </Link>
+        <LoginForm />
+        <Footer theme="light" />
+      </CenteredContainer>
     </div>
   );
 }
