+++ lex_main/src/components/authentication/Signup.jsx
@@ -5,11 +5,11 @@
  */
 
 import React from "react";
-import { Container } from "react-bootstrap";
 import { Link } from "react-router-dom";
 import "../../index.css";
 
 import Footer from "../Footer";
+import CenteredContainer from "./CenteredContainer";
 import SignupForm from "./SignupForm";
 
 function Signup() {
@@ -19,18 +19,13 @@ function Signup() {
 
   return (
     <div className="gradient-background">
-      <Container
-        className="d-flex align-items-center justify-content-center"
-        style={{ minHeight: "100vh" }}
-      >
-        <div className="w-100" style={{ maxWidth: "400px" }}>
-          <Link to="/landing" style={{ textDecoration: "none" }}>
-            <h1 className="logo logo-big logo-white text-center">LEX</h1>
-          </Link>
-          <SignupForm />
-          <Footer theme="light" />
-        </div>
-      </Container>
+      <CenteredContainer>
+        <Link to="/landing" style={{ textDecoration: "none" }}>
+          <h1 className="logo logo-big logo-white text-center">LEX</h1>
+        </Link>
+        <SignupForm />
+        <Footer theme="light" />
+      </CenteredContainer>
     </div>
   );
 }
